+++ Changelog.md
@@ -6,6 +6,7 @@
 - Added CRUD for Countries and Provinces
 - Added the `globe` icon
 - Added support for setting `empty.html` on the table widget
+- Added Customer purchases chart to the Customer show page
 
 ## 4.3.0
 ##### 2024-10-29